+++ src/Framework/Bootstrap/Setup/SetupMerger.php
@@ -34,6 +34,7 @@ final class SetupMerger
         $this->mergeAliases($other);
         $this->mergeContextualBindings($other);
         $this->mergeHandlerRegistries($other);
+        $this->mergeLazyServices($other);
         $this->mergePlugins($other);
         $this->mergeGacelaConfigsToExtend($other);
 
@@ -158,6 +159,13 @@ final class SetupMerger
     {
         if ($other->isPropertyChanged(SetupGacela::handlerRegistries)) {
             $this->original->mergeHandlerRegistries($other->getHandlerRegistries());
+        }
+    }
+
+    private function mergeLazyServices(SetupGacela $other): void
+    {
+        if ($other->isPropertyChanged(SetupGacela::lazyServices)) {
+            $this->original->mergeLazyServices($other->getLazyServices());
         }
     }
 }